Public Mesh drawing API and sample This adds methods to the ARVisualizer API to draw triangles and quads, and adds a sample application to demonstrate use of them.
These APIs are not finished, yet, and are subject to change, but the basic
idea is that an application will make a call to Add<shape> with the appropriate
arguments to create a specific shape, and will get back an opaque handle that
can later be used to Update, Remove, etc. the previously-added shape.

Also fixes a bug in rendering where we weren't passing the correct index
offset to glDrawElementsBaseVertex.