Better Thread Safety GLFW Requires that many calls are always made from the same thread. To avoid requiring that all applications using ARVisualizer take care to make certain calls from certain threads, this adds a WindowManager class, which handles deferring all the necessary calls to GLFW to one specific thread, regardless of which thread is making the request.
WindowManager is implemented as a Singleton, and creates a new thread upon instantiation.
It then provides several public methods for submitting work to be done on this new thread.
Any call which results in the creation of new objects on the background thread will block
until the background thread is finished, so users of WindowManager can treat it as though
it has a synchronous API.