Allow deletion of mesh objects This extends the Renderer to actually generate unique handles for new meshes, and to be able to delete meshes when given the correct handle. If an unassigned or invalid handle is given, the Renderer will do nothing.
The implementation for the actual removal process is a bit naive. Any time one
or more meshes is deleted, we throw out all of the data in the VertexBuffer that
mesh was assigned to and re-add all the remaining meshes. This ensures any meshes
stored at a higher index in the buffer get new correct offsets.

It may be better in the future to remove just the region of the buffer corresponding
to the removed mesh, then manually update the indices for all the meshes that were
stored at a higher index than the removed mesh.