Prevent race condition when adding new meshes It was possible for the renderer to try to render a mesh which was incomplete due to a race condition caused by using vector::push_back without a mechanism to ensure the push was finished before the renderer tried to access the result.
This adds a separate buffer for new meshes, which the renderer copies into
the buffer of meshes used for rendering only when no other thread has a lock
on the mutex. This may delay rendering of some meshes by a frame or two if the
renderer isn't able to take the lock immediately after a mesh is added.