Make calls to glViewPort thread-safe This is a workaround for glfw's event handling. glfwPollEvents will invoke all registered event handlers at the time it's called, which can cause one Renderer to invoke the event handlers from another if we have multiple windows open. This is only really an issue in cases where we access thread-specific resources, like the GL Context, and was causing glViewPort to sometimes be called from the wrong thread.
Officially, we should be calling glfwPollEvents from the WindowManger's thread, then either stealing the
GL context from the Renderer which needs to call glViewPort in order to make the call from the WindowManger
thread, or sending a message to the renderer thread to do it (via the CommandQueue).

Both solutions are likely slower and/or more expensive than just calling glViewPort each frame, and the other
events aren't affected.